requirements.txt

atlassian-python-api
loguru
PyYAML
SQLAlchemy
flask
pynvml
pyopenproject
sqlalchemy_utils
pydantic


sentence-transformers==2.2.2
transformers==4.38.2
accelerate==0.27.2
bitsandbytes==0.41.3.post2
tokenizers==0.15.2
flask==3.0.2
mlserver==1.5.0
flask-cors==4.0.0
pandas==2.2.1
chromadb==0.4.24
langchain==0.0.331
pydantic==1.10.10
openai==0.28.1
